@@ -2,7 +2,45 @@
     ViewData["Title"] = "Home Page";
 }
 
-<div class="text-center">
-    <h1 class="display-4">Welcomxzczxe</h1>
-    <p>Learn about <a href="https://docs.microsoft.com/aspnet/core">building Web apps with ASP.NET Core</a>.</p>
-</div>
+
+    <div class="container p-5  d-flex justify-content-center">
+        <table class="table">
+            <thead class="thead-dark">
+                <tr>
+                    <th scope="col">#</th>
+                    <th scope="col">Başlık</th>
+                    <th scope="col">Oluşturma Tarihi</th>
+                    <th scope="col">İşlemler</th>
+                </tr>
+            </thead>
+            <tbody>
+                <tr>
+                    <th scope="row">1</th>
+                    <td>Mark</td>
+                    <td>Otto</td>
+                    <td>
+                        <a href="#" class="btn-sm btn-info fas fa-clipboard-list" title="Önizleme" />
+                        <a href="#" class="btn-sm btn-danger fas fa-trash" title="Sil" />
+                    </td>
+                </tr>
+                <tr>
+                    <th scope="row">2</th>
+                    <td>Jacob</td>
+                    <td>Thornton</td>
+                    <td>
+                        <a href="#" class="btn-sm btn-info fas fa-clipboard-list" title="Önizleme" />
+                        <a href="#" class="btn-sm btn-danger fas fa-trash" title="Sil" />
+                    </td>
+                </tr>
+                <tr>
+                    <th scope="row">3</th>
+                    <td>Larry</td>
+                    <td>the Bird</td>
+                    <td>
+                        <a href="#" class="btn-sm btn-info fas fa-clipboard-list" title="Önizleme" />
+                        <a href="#" class="btn-sm btn-danger fas fa-trash" title="Sil" />
+                    </td>
+                </tr>
+            </tbody>
+        </table>
+        </div>
